Move ConcatSource to iobase.
